minor: used interpret_cast in server to fit cpplints' preferences
