+++ includes/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: mbernard <mbernard@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 11:50:56 by faboussa          #+#    #+#             */
-/*   Updated: 2024/10/29 10:56:05 by mbernard         ###   ########.fr       */
+/*   Updated: 2024/10/29 12:21:27 by mbernard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,6 +50,7 @@ enum Command {
   PING,
   CAP,
   USER,
+  PASS,
   UNKNOWN
 };
