+++ src/Server.cpp
@@ -6,7 +6,7 @@
 /*   By: mbernard <mbernard@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 11:50:56 by faboussa          #+#    #+#             */
-/*   Updated: 2024/10/29 11:02:18 by mbernard         ###   ########.fr       */
+/*   Updated: 2024/10/29 15:17:45 by mbernard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -182,11 +182,9 @@ void Server::handleOtherMessage(Client &client, const std::string &message) {
       // ERR_UNKNOWNCOMMAND (421)
       std::cerr << RED "Unknown command" RESET << std::endl;
       continue;
-    }
-    else if (cmd == CAP || cmd == USER) {
+    } else if (cmd == CAP || cmd == USER) {
       continue;
-    }
-    else {
+    } else {
       handleCommand(command, argument, client.getFd());
     }
   }
@@ -241,14 +239,8 @@ void Server::handleClientMessage(int fd) {
   if (client.isAccepted() == false) {
     handleInitialMessage(client, message);
   } else {
-    std::istringstream iss(message);
-    std::string command;
-    std::string argument;
-    iss >> command;
-    if (command == "JOIN")
-      handleCommand(command, argument, fd);
-    else
-      sendToAllClients(message);
+    // sendToAllClients(message);
+    handleOtherMessage(client, message);
   }
 }
 
@@ -258,7 +250,8 @@ void Server::acceptNewClient() {
   struct pollfd newPoll;
   socklen_t len = sizeof(cliadd);
 
-  int newClientFd = accept(_socketFd, (sockaddr *)&cliadd, &len);
+  int newClientFd = accept(_socketFd,
+                          reinterpret_cast<sockaddr *>(&cliadd), &len);
   if (newClientFd == -1) {
     std::cerr << RED "Failed to accept new client" RESET << std::endl;
     return;
@@ -316,12 +309,7 @@ void Server::handleCommand(const std::string &command,
                            const std::string &argument, int fd) {
   if (command.empty()) return;
   if (command == "JOIN") {
-    // std::string channelName;
-    // command >> channelName;
-    // if (_channels.find(channelName) == _channels.end()) {
-    //   _channels[channelName] = Channel(channelName);
-    // }
-    // _channels[channelName].acceptClientInTheChannel(_clients[fd]);
+    // join
   } else if (command == "KICK") {
     // Exclure un client du canal
   } else if (command == "INVITE") {
@@ -342,26 +330,12 @@ void Server::handleCommand(const std::string &command,
     // Déconnecter le client
   } else if (command == "PING") {
     // client.sendNumericReply(1, "PONG");
+  } else if (command == "PASS" || command == "USER") {
+    // if (argument.empty())
+    //   sendNumericReply(461, ERR_NEEDMOREPARAMS);
+    // else
+    //   sendNumericReply(462, ERR_ALREADYREGISTERED);
   } else {
     // Commande inconnue
   }
 }
-
-// void Server::handlePassword(int fd) {
-//   char buffer[1024] = {0};
-//   std::memset(buffer, 0, sizeof(buffer));
-//   int valread = recv(fd, buffer, sizeof(buffer), 0);
-
-//   switch (valread) {
-//     case -1:
-//       std::cerr << RED "Error while receiving message" RESET << std::endl;
-//       // fallthrough
-//     case 0:
-//       clearClient(fd);
-//       return;
-//   }
-//   std::cout << YELLOW << buffer << RESET << std::endl;
-//   std::string message(buffer, valread);
-//   std::istringstream iss(message);
-//   sendToAllClients(message);
-// }
